Refactors and remove the code
