+++ app/src/main/java/com/udacity/googleindiascholarships/challenges/ui/ChallengesFragment.java
@@ -1,4 +1,4 @@
-package com.udacity.googleindiascholarships;
+package com.udacity.googleindiascholarships.challenges.ui;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
@@ -6,21 +6,24 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.udacity.googleindiascholarships.R;
+
 /**
  * Created by jha.anuj.2108 on 13-04-2018.
  */
 
-public class ChallengesFragment extends android.support.v4.app.Fragment{
+public class ChallengesFragment extends android.support.v4.app.Fragment {
+
     @Nullable
     @Override
-    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container,
+                             @Nullable Bundle savedInstanceState) {
         return inflater.inflate(R.layout.fragment_challenges,container,false);
     }
 
     @Override
     public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
-
         getActivity().setTitle("Challenges");
     }
 }
