+++ app/src/main/java/com/udacity/googleindiascholarships/members/ui/MembersFragment.java
@@ -1,4 +1,4 @@
-package com.udacity.googleindiascholarships;
+package com.udacity.googleindiascholarships.members.ui;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
@@ -6,14 +6,18 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.udacity.googleindiascholarships.R;
+
 /**
  * Created by jha.anuj.2108 on 13-04-2018.
  */
 
 public class MembersFragment extends android.support.v4.app.Fragment{
+
     @Nullable
     @Override
-    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container,
+                             @Nullable Bundle savedInstanceState) {
         View rootView = inflater.inflate(R.layout.fragment_members, container, false);
         return rootView;
     }
