+++ app/src/main/res/layout/fragment_projects.xml
@@ -9,7 +9,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_gravity="center"
-        android:text="Challenges"
+        android:text="Project"
         android:textColor="@android:color/black"
         android:textSize="24sp"
         android:textStyle="bold" />
