+++ app/src/main/res/values/colors.xml
@@ -3,6 +3,5 @@
     <color name="colorPrimary">#00aeef</color>
     <color name="colorPrimaryDark">#0298d1</color>
     <color name="colorAccent">#24c2fd</color>
-
-
+    <color name="navBarBackground">#ffffff</color>
 </resources>
